New SM code cortos compatible

+++ CoRTOS/bubblesort.h
@@ -1,12 +1,12 @@
 //#include <stdio.h>
-#include "cortos.h"
+#include "CORTOS.h"
 #include <stdlib.h>
 #include <math.h>
 
-void bubbleSort(double arr[][3], int n)
+void bubbleSort(long double arr[][3], int n)
 {
     int i, j;
-    double temp[3];
+    long double temp[3];
     for (i = 0; i < n; i++)
     {
         for (j = 0; j < n - i - 1; j++)
